warnings cleanup

+++ src/libawkward/forth/ForthOutputBuffer.cpp
@@ -363,7 +363,7 @@ namespace awkward {
   ForthOutputBufferOf<OUT>::write_one_string(char* string_buffer, int64_t length) noexcept {
     int64_t next = length_ + length;
     maybe_resize(next);
-    std::memcpy(&ptr_.get()[length_], string_buffer, length);
+    std::memcpy(&ptr_.get()[length_], string_buffer, (size_t)length);
     length_ = next;
   }
